refactor: add situations as a module
