+++ src/modules/situations/SituationMessageBox.tsx
