+++ src/modules/situations/SituationSectionItem.tsx
