+++ src/modules/situations/utils.ts
